move to new examples dir
